requirements.txt
# --- API server ---
flask==3.0.3
python-dotenv==1.0.1

# --- OpenAI ---
openai==1.40.6

# --- RAG / Vector search ---
faiss-cpu==1.8.0.post1
numpy==1.26.4

# --- Data / text ---
pandas==2.2.2
nltk==3.9.1

# --- Utilities ---
pydantic==2.8.2
tqdm==4.66.5

streamlit==1.37.1
requests==2.32.3

fastapi==0.115.0
uvicorn==0.30.6